pyproject.toml
[tool.poetry]
name = "lessons"
version = "0.1.0"
description = ""
authors = ["Sergey Sizov <sssss_aaaaa_2011@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
customtkinter = "^5.2.2"
tkintermapview = "^1.29"
flet = "==0.28.0.dev4456"
dearpygui = "==2.0.0"
tkcalendar = "^1.6.1"
ctkmessagebox = "==2.7"
plyer = "^2.1.0"
niquests = "^3.14.0"
anyio = "==4.9.0"
certifi = "==2025.1.31"
charset-normalizer = "==3.4.1"
click = "==8.1.8"
fastapi = "==0.115.12"
decorator = "==5.2.1"
idna = "==3.10"
httptools = "==0.6.4"
pydantic-core = "==2.33.1"
jinja2 = "==3.1.6"
pydantic = "==2.11.3"
markupsafe = "==3.0.2"
pillow = "==11.2.1"
rich = "==14.0.0"
urllib3 = "==2.4.0"
websockets = "==15.0.1"
six = "==1.17.0"
pywin32 = "==310"
python-dotenv = "==1.1.0"
pyyaml = "==6.0.2"
pygments = "==2.19.1"
requests = "==2.32.3"
urllib3-future = "==2.12.917"
xmltodict = "^0.14.2"
scrapy = "^2.12.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
exif = "^1.6.1"
pycountry = "^24.6.1"
ffmpeg-python = "^0.2.0"
reverse-geocoder = "^1.5.1"
aiogram = "^3.20.0.post0"
yadisk = "^3.3.0"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
